Separate CodeGenerator from its impl
