+++ crates/mockiato-codegen/src/code_generator_impl/bound_lifetimes.rs
