+++ crates/mockiato-codegen/src/code_generator_impl/lifetime_rewriter.rs
